requirements.txt
playwright
plotly
pystache
markdownify
openai
python-dotenv